del survey, fix usecase order

+++ internal/pkg/order/usecase/usecase_test.go
@@ -41,7 +41,7 @@ func TestOrderUsecase_CreateOrder(t *testing.T) {
 			prepare: func(f *fields) {
 				f.repoAddress.EXPECT().ReadCurrentAddress(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Address{Id: uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")}, nil)
 				f.repoCart.EXPECT().ReadCart(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Cart{}, nil)
-				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), 1).Return(models.Order{}, nil)
+				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), int64(1)).Return(models.Order{}, nil)
 				f.repoCart.EXPECT().DeleteCart(gomock.Any(), gomock.Any()).Return(nil)
 			},
 			args:    args{context.Background(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")},
@@ -70,7 +70,7 @@ func TestOrderUsecase_CreateOrder(t *testing.T) {
 			prepare: func(f *fields) {
 				f.repoAddress.EXPECT().ReadCurrentAddress(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Address{Id: uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")}, nil)
 				f.repoCart.EXPECT().ReadCart(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Cart{}, nil)
-				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), 1).Return(models.Order{}, errors.New("dummyError"))
+				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), int64(1)).Return(models.Order{}, errors.New("dummyError"))
 			},
 			args:    args{context.Background(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")},
 			wantErr: true,
@@ -80,7 +80,7 @@ func TestOrderUsecase_CreateOrder(t *testing.T) {
 			prepare: func(f *fields) {
 				f.repoAddress.EXPECT().ReadCurrentAddress(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Address{Id: uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")}, nil)
 				f.repoCart.EXPECT().ReadCart(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Cart{}, nil)
-				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), 1).Return(models.Order{}, nil)
+				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), int64(1)).Return(models.Order{}, nil)
 				f.repoCart.EXPECT().DeleteCart(gomock.Any(), gomock.Any()).Return(errors.New("dummyError"))
 			},
 			args:    args{context.Background(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")},
@@ -108,7 +108,7 @@ func TestOrderUsecase_CreateOrder(t *testing.T) {
 			prepare: func(f *fields) {
 				f.repoAddress.EXPECT().ReadCurrentAddress(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Address{Id: uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")}, nil)
 				f.repoCart.EXPECT().ReadCart(gomock.Any(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")).Return(models.Cart{}, nil)
-				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), 1).Return(models.Order{}, orderRepo.ErrPoductNotFound)
+				f.repoOrder.EXPECT().CreateOrder(gomock.Any(), models.Cart{}, uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a"), int64(1)).Return(models.Order{}, orderRepo.ErrPoductNotFound)
 			},
 			args:    args{context.Background(), uuid.FromStringOrNil("fb11fe90-09bb-4e72-98a5-5ffba93aa39a")},
 			wantErr: true,
